+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version (development version)
+
+ * ...
+
+
 # Version 0.2.2 [2023-05-25]
 
 ## New Features